+++ src/Model/Utils/Algorithm.cpp
@@ -12,7 +12,7 @@ void Algorithm::SetTotalEpoch(unsigned int total_epoch)
 
 float Algorithm::LearningRate(float learning_rate, unsigned int actual_epoch)
 {
-	return learning_rate * std::exp(- static_cast<int>(actual_epoch) / static_cast<int>(total_epoch_));
+	return static_cast<float>(learning_rate * std::exp(- static_cast<int>(actual_epoch) / static_cast<int>(total_epoch_)));
 }
 
 float Algorithm::Radius(float sigma, unsigned int current_epoch, float time_constant)
